Structure changes
Structure changes in js files
