+++ src/auth/src/main/resources/static/js/listeners/login.js
@@ -1,10 +1,10 @@
 "use strict";
 
-import { userService } from "../auth-services/services.js";
+import { userAuthService } from "../auth-services/services.js";
 import { DOM_DATA_GETTERS } from "../utils/dom-parser.js";
 
 const names = ["username", "passwd"];
 
 DOM_DATA_GETTERS.id("login-user").addEventListener("click", () => {
-  userService.login(...names);
+  userAuthService.login(...names);
 });
