+++ src/auth/src/main/resources/static/js/listeners/logout.js
@@ -1,10 +1,10 @@
 "use strict";
 
-import { userService } from "../auth-services/services.js";
+import { userAuthService } from "../auth-services/services.js";
 import { DOM_DATA_GETTERS } from "../utils/dom-parser.js";
 
 DOM_DATA_GETTERS.id("logout-user").addEventListener("click", () => {
-  userService.logout({});
+  userAuthService.logout({});
 });
