+++ src/auth/src/main/resources/static/js/listeners/register.js
@@ -1,10 +1,10 @@
 "use strict";
 
-import { userService } from "../auth-services/services.js";
+import { userAuthService } from "../auth-services/services.js";
 import { DOM_DATA_GETTERS } from "../utils/dom-parser.js";
 
 const names = ["username", "email", "passwd"];
 
 DOM_DATA_GETTERS.id("register-user").addEventListener("click", () => {
-  userService.register(...names);
+  userAuthService.register(...names);
 });
