+++ src/auth/src/main/resources/static/js/services/abstract-service/abstract-service.js
