+++ src/auth/src/main/resources/static/js/services/auth-services/user-auth-service-class.js
