+++ src/auth/src/main/resources/static/js/services/callbacks/email-verify-success.js
@@ -1,9 +1,9 @@
 "use strict";
 
-import get from "../methods/fetch-get.js";
+import { get } from "../../utils/fetch/methods/fetch-methods.js";
 import ROUTES from "../../data/routes.js";
 
-const emailVerifySuccess = () => {
+const emailVerifySuccess = (...args) => {
   const REDIRECT_ROUTE = ROUTES.resetPassword;
   get({ route: REDIRECT_ROUTE });
 }
