+++ src/auth/src/main/resources/static/js/services/callbacks/generic-msg-fail.js
