+++ src/auth/src/main/resources/static/js/services/reset-service/reset-service-class.js
