+++ src/auth/src/main/resources/static/js/services/services.js
@@ -1,0 +1,6 @@
+"use strict";
+
+import userAuthService from "./auth-services/user-auth-service.js";
+import resetService from "./reset-service/reset-service.js";
+
+export { userAuthService, resetService };
