+++ src/auth/src/main/resources/static/js/utils/fetch/fetch-consts.js
