+++ src/auth/src/main/resources/static/js/utils/fetch/jwt/jwt-fetch-methods.js
