+++ src/auth/src/main/resources/static/js/utils/fetch/jwt/jwt-get.js
