+++ src/auth/src/main/resources/static/js/utils/fetch/methods/fetch-methods.js
