+++ src/auth/src/main/resources/static/js/utils/parser/dom-parser.js
